Equipping A Weapon
I am working on the equipping a weapon. I have got the position set up, I am working on getting the animations in for the different weapons. Though this is a bit difficult since the teacher didn't go through it in class and the lecture notes are not complete enough to use them to work on it. I might just keep going and skip the animations for now.