Added keybindings, resolution options, windowing options and quality options
I am working on the options menu currently. I got the keybindings to work, so that the player can bind the movement keys to any key on the keyboard. I also got in resolution setting. I think it will work but I haven't tested it in build yet. Next I am going to work on adding audio and audio options